Fix FP16/FP32 dtype mismatch in DeltaNet state update
beta was FP32 (sigmoid of FP32 dense output), causing state to be
promoted to FP32 after first iteration. Cast beta to FP16 so the
entire sequential scan stays in FP16.

+++ experiments/exp1_thinker/thinker15b/train.py
@@ -563,7 +563,7 @@ class DeltaNetCell(nn.Cell):
         x_f32 = x.astype(ms.float32)
         K = self.key_proj(x_f32).astype(ms.float16)   # (B, S, D)
         V = self.value_proj(x_f32).astype(ms.float16)  # (B, S, D)
-        beta = ops.sigmoid(self.beta_proj(x_f32))      # (B, S, NH)
+        beta = ops.sigmoid(self.beta_proj(x_f32)).astype(ms.float16)  # (B, S, NH)
 
         # Initialize state: broadcast to batch
         state = ops.Tile()(self.initial_state, (B, 1, 1, 1))  # (B, NH, HD, HD)
